Remove apt-get-install dep
The dependency introduces a new layer, only got used in the entrypoint
script (and not in the Dockerfile) and also its DEBIAN_FRONTEND env
variable is falsely implemented.

As of today, it uses
DEBIAN_FRONTEND=noninteractive apt-get update && apt-get install ...
which will only export the DEBIAN_FRONTEND to the update call, but not
to the install call.

Signed-off-by: Benedikt Heine <bebe@bebehei.de>