Remove APT_PROXY argument
The apt proxy introduced a new layer of caching outside of docker.

This may be neat, but actually the docker caching layer should suffice.
It shouldn't be necessary to rebuild the package layer regularly.

It may be the case, that the cached package layer gets invalidated.
And for this, we should rearrange the layer structure.

Signed-off-by: Benedikt Heine <bebe@bebehei.de>