Improve ENV and LABEL usage
Define all env variables as ENV var=value to make chaining easier and
put them in one layer to reduce layer size.

The DEBIAN_FRONTEND should get applied on a per-call basis and not as a
global env. [0]

Also the ENV layers should be on the lowest possible position,
especially the LC_ALL variable to make use of docker's cache.

[0]: https://github.com/moby/moby/issues/4032#issuecomment-34597177

Signed-off-by: Benedikt Heine <bebe@bebehei.de>